+++ pyspectrum/peak_identification/__init__.py
@@ -1,5 +1,6 @@
 from .peaks_fit import GaussianWithBGFitting
-from .find_gamma_peaks import FindPeaks, Convolution
+from .convolution import Convolution
+from .find_gamma_peaks import FindPeaks
 from .zero_area_functions import gaussian_2_dev, asymmetrical_rect_zero_area
 from .peak import Peak
